+++ medacy/pipelines/lstm_systematic_review_pipeline.py
@@ -1,10 +1,10 @@
 import spacy
 
-from medacy.pipelines.base.base_pipeline import BasePipeline
-from medacy.pipeline_components.feature_overlayers.gold_annotator_component import GoldAnnotatorComponent
 from medacy.pipeline_components.feature_extracters.discrete_feature_extractor import FeatureExtractor
+from medacy.pipeline_components.feature_overlayers.gold_annotator_component import GoldAnnotatorComponent
 from medacy.pipeline_components.learners.bilstm_crf_learner import BiLstmCrfLearner
 from medacy.pipeline_components.tokenizers.systematic_review_tokenizer import SystematicReviewTokenizer
+from medacy.pipelines.base.base_pipeline import BasePipeline
 
 
 class LstmSystematicReviewPipeline(BasePipeline):
