+++ medacy/pipelines/testing_pipeline.py
@@ -1,10 +1,10 @@
 import sklearn_crfsuite
 import spacy
 
-from medacy.pipelines.base.base_pipeline import BasePipeline
-from medacy.pipeline_components.feature_overlayers.gold_annotator_component import GoldAnnotatorComponent
 from medacy.pipeline_components.feature_extracters.discrete_feature_extractor import FeatureExtractor
+from medacy.pipeline_components.feature_overlayers.gold_annotator_component import GoldAnnotatorComponent
 from medacy.pipeline_components.tokenizers.clinical_tokenizer import ClinicalTokenizer
+from medacy.pipelines.base.base_pipeline import BasePipeline
 
 
 class TestingPipeline(BasePipeline):
